use home page layout component
