+++ layout/pages/home/style.scss
